fix: the library builds correctly now
