+++ package.json
@@ -6,10 +6,9 @@
   "license": "MIT",
   "sideEffects": false,
   "main": "dist/module.js",
-  "typings": "./types/index.d.ts",
+  "typings": "dist/types",
   "files": [
-    "dist",
-    "types"
+    "dist"
   ],
   "scripts": {
     "build": "siroc build && yarn build:runtime",
