--- src/types/module.ts
@@ -1,9 +1,0 @@
-import { CookieSerializeOptions } from 'cookie'
-
-export interface ModuleOptions {
-  url: string
-  entities: string[],
-  key: string,
-  expires: 'session' | string | number,
-  cookie: CookieSerializeOptions
-}
